feat: add test

+++ packages/car-management-api/src/controllers/admin/admin-user.controller.ts
@@ -1,21 +1,47 @@
+import type { OmitPasswordHash } from "@/types/typeHelper";
+import type { AdminUser } from "@prisma/client";
 import type { Request, Response } from "express";
 import * as adminUserService from "../../services/admin-user.service";
 
-// GET /api/admin/users
+export const hasAdminManipulationPermission = (
+  user: OmitPasswordHash<AdminUser>,
+  targetUser: OmitPasswordHash<AdminUser>,
+): boolean => {
+  switch (user.role) {
+    case "super_admin":
+      return true;
+    case "admin":
+      return targetUser.role !== "super_admin";
+    case "tenant_admin":
+      return targetUser.role !== "super_admin" && targetUser.role !== "admin" && targetUser.tenantId === user.tenantId;
+    case "tenant_viewer":
+      return false;
+    default:
+      return false;
+  }
+};
+
+// GET /api/v1/admin/users
 export const getAllAdminUsers = async (req: Request, res: Response) => {
   try {
-    const users = await adminUserService.getAllAdminUsers();
+    const users = (await adminUserService.getAllAdminUsers()).filter((user) =>
+      hasAdminManipulationPermission(req.user!, user),
+    );
     res.json(users);
   } catch (error) {
     res.status(500).json({ message: "Error fetching admin users" });
   }
 };
 
-// GET /api/admin/users/:id
+// GET /api/v1/admin/users/:id
 export const getAdminUserById = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
     const user = await adminUserService.getAdminUserById(id);
+    if (hasAdminManipulationPermission(req.user!, user!)) {
+      res.status(403).json({ message: "Forbidden" });
+      return;
+    }
     if (user) {
       res.json(user);
     } else {
@@ -26,14 +52,11 @@ export const getAdminUserById = async (req: Request, res: Response) => {
   }
 };
 
-// POST /api/admin/users
+// POST /api/v1/admin/users
 export const createAdminUser = async (req: Request, res: Response) => {
   try {
     const { user } = req;
-    if (
-      adminUserService.getAdminUserRoleLevel(user!.role) <
-      adminUserService.getAdminUserRoleLevel(req.body.role)
-    ) {
+    if (!hasAdminManipulationPermission(user!, req.body)) {
       res.status(403).json({ message: "Forbidden" });
       return;
     }
@@ -44,10 +67,15 @@ export const createAdminUser = async (req: Request, res: Response) => {
   }
 };
 
-// PUT /api/admin/users/:id
+// PUT /api/v1/admin/users/:id
 export const updateAdminUser = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const { user } = req;
+    if (!hasAdminManipulationPermission(user!, req.body)) {
+      res.status(403).json({ message: "Forbidden" });
+      return;
+    }
     const updatedUser = await adminUserService.updateAdminUser(id, req.body);
     if (updatedUser) {
       res.json(updatedUser);
@@ -59,10 +87,15 @@ export const updateAdminUser = async (req: Request, res: Response) => {
   }
 };
 
-// DELETE /api/admin/users/:id
+// DELETE /api/v1/admin/users/:id
 export const deleteAdminUser = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const { user } = req;
+    if (!hasAdminManipulationPermission(user!, req.body)) {
+      res.status(403).json({ message: "Forbidden" });
+      return;
+    }
     await adminUserService.deleteAdminUser(id);
     res.status(204).send();
   } catch (error) {
